src/requirements.txt
psycopg[binary]
alembic
SQLAlchemy

fastapi[all]
strawberry-graphql
strawberry-sqlalchemy-mapper

python-jose[pycryptodome]